+++ src/components/datagrid/BodyRow.js
@@ -6,14 +6,10 @@ import Row from './Row';
 import cx from 'classnames';
 import theme from './theme.css';
 
-const allowedParentNodes = ['datagrid-body-row', 'datagrid-cell'];
-
 class BodyRow extends PureComponent {
   handleClick = (event) => {
-    if (allowedParentNodes.includes(event.target.parentNode.dataset.teamleaderUi)) {
-      const { onClick } = this.props;
-      onClick && onClick(event);
-    }
+    const { onClick } = this.props;
+    onClick && onClick(event);
   };
 
   render() {
@@ -50,7 +46,7 @@ class BodyRow extends PureComponent {
         {...others}
       >
         {selectable && (
-          <Cell align="center" flex="min-width">
+          <Cell align="center" flex="min-width" onClick={(event) => event.stopPropagation()}>
             <Checkbox checked={selected} onChange={onSelectionChange} size={checkboxSize} />
           </Cell>
         )}