+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Removed
 
+- [Breaking] `StatusLabel`: remove the hardcoded horizontal margin. You can optionally pass the `marginHorizontal` prop with a corresponding value. ([@driesd](https://github.com/driesd) in [#970](https://github.com/teamleadercrm/ui/pull/970)
+
 ### Fixed
 
 ### Dependency updates